HGI-9643: Move utilities to their respective files and merge with existing functionalities
