+++ gluestick/utils/exceptions.py
@@ -1,0 +1,4 @@
+class CustomValidationError(Exception):
+    def __init__(self, error):
+        super().__init__(error)
+        self.error = error
